pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "unykornx-hydragrid"
version = "0.0.1"
description = "UnykornX HydraGrid: Multi-asset AI trading system (equities + energy)"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "HydraGrid Dev Team"}
]
requires-python = ">=3.10"

dependencies = [
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "pyarrow>=12.0",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.1",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
    "ruff>=0.1",
]
engine = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scipy>=1.10",
]
data = [
    "pandas>=2.0",
    "pyarrow>=12.0",
    "pandas-datareader>=0.10",
    "yfinance>=0.2",
]
api = [
    "fastapi>=0.104",
    "uvicorn>=0.24",
    "pydantic>=2.0",
]
swarm = [
    "anthropic>=0.7",
]
onchain = [
    "web3>=6.0",
    "eth-account>=0.9",
    "eth-abi>=4.2",
]
all = [
    "unykornx-hydragrid[dev,engine,data,api,swarm,onchain]",
]

[tool.setuptools.packages.find]
where = [
    "packages/engine-core",
    "packages/data-pipelines",
    "packages/strategies-stock",
    "packages/strategies-energy",
    "packages/swarm",
    "packages/risk",
    "packages/analytics",
    "packages/infra"
]
include = ["unykornx*", "hydragrid*"]

[tool.black]
line-length = 120
target-version = ["py310", "py311", "py312"]
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 120
multi_line_mode = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_any_generics = false
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E203", "W503"]
